requirements.txt
# ── Web framework ────────────────────────────────
fastapi>=0.111
uvicorn[standard]>=0.29

# ── Config ───────────────────────────────────────
pydantic>=2.7
pydantic-settings>=2.3
python-dotenv>=1.0

# ── PostgreSQL ────────────────────────────────────
psycopg2-binary>=2.9
pgvector>=0.3

# ── Neo4j ─────────────────────────────────────────
neo4j>=5.20

# ── Object Storage ────────────────────────────────
minio>=7.2

# ── Crawler ───────────────────────────────────────
httpx>=0.27
trafilatura>=1.10
readability-lxml>=0.8

# ── NLP / Extraction ──────────────────────────────
pyyaml>=6.0
simhash>=2.1

# ── Retry / Logging ───────────────────────────────
tenacity>=8.3

# ── Dev / Testing ─────────────────────────────────
pytest>=8.2
pytest-asyncio>=0.23
httpx>=0.27   # also used for test client